+++ src/JustCSharp.ShopeeSDK/Domain/Order/CompactOrderData.cs
@@ -239,7 +239,7 @@ public class CompactOrderData
         public List<int> RatingCount { get; set; }
 
         [JsonPropertyName("rating_star")]
-        public int RatingStar { get; set; }
+        public decimal RatingStar { get; set; }
 
         [JsonPropertyName("shop_id")]
         public int ShopId { get; set; }
