+++ src/JustCSharp.ShopeeSDK/Domain/Order/Order.cs
@@ -260,7 +260,7 @@ public class Order
         public List<int> RatingCount { get; set; }
 
         [JsonPropertyName("rating_star")]
-        public int RatingStar { get; set; }
+        public decimal RatingStar { get; set; }
 
         [JsonPropertyName("shop_id")]
         public int ShopId { get; set; }
